requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
openai
mangum
pydantic
